+++ weblate/api/serializers.py
@@ -2042,6 +2042,7 @@ class ComponentSerializer(RemovableSerializer[Component]):
         "new_base",
         "file_format",
         "repoweb",
+        "repoweb_translations",
         "merge_style",
         "auto_lock_error",
         "language_regex",
@@ -2224,6 +2225,7 @@ class ComponentSerializer(RemovableSerializer[Component]):
             "enforced_checks",
             "restricted",
             "repoweb",
+            "repoweb_translations",
             "report_source_bugs",
             "merge_style",
             "commit_message",
@@ -2319,6 +2321,7 @@ class ComponentSerializer(RemovableSerializer[Component]):
             result["git_export"] = None
             result["push_branch"] = None
             result["repoweb"] = None
+            result["repoweb_translations"] = None
             result["linked_component"] = None
         return result