+++ weblate/trans/alerts/config.py
@@ -238,17 +238,17 @@ class BrokenBrowserURL(BaseAlert):
 
     @staticmethod
     def check_component(component: Component) -> bool | dict | None:
-        location_error = None
-        location_link = None
-        if component.repoweb:
-            if component.source_language_id is None:
-                return False
-            translation = (
-                component.translation_set.filter(unit__state__gte=STATE_TRANSLATED)
-                .exclude(language_id=component.source_language_id)
-                .first()
-            )
+        if not component.repoweb and not component.repoweb_translations:
+            return False
+        if component.source_language_id is None:
+            return False
+        translation = (
+            component.translation_set.filter(unit__state__gte=STATE_TRANSLATED)
+            .exclude(language_id=component.source_language_id)
+            .first()
+        )
 
+        if component.repoweb:
             if translation:
                 allunits = translation.unit_set
             else:
@@ -264,9 +264,30 @@ class BrokenBrowserURL(BaseAlert):
                         location_link,
                         validators=(WeblateURLValidator(),),
                     )
+                    if location_error:
+                        return {"link": location_link, "error": location_error}
                     break
-        if location_error:
-            return {"link": location_link, "error": location_error}
+
+        browser_translation = translation
+        if component.repoweb_translations and browser_translation is None:
+            browser_translation = (
+                component.translation_set.exclude(
+                    language_id=component.source_language_id
+                )
+                .exclude(filename="")
+                .first()
+            )
+        if component.repoweb_translations and browser_translation:
+            location_link = component.get_repoweb_link(
+                browser_translation.filename, "1", is_translation=True
+            )
+            if location_link is not None:
+                location_error = _get_validated_uri_error(
+                    location_link,
+                    validators=(WeblateURLValidator(),),
+                )
+                if location_error:
+                    return {"link": location_link, "error": location_error}
         return False